Major refactors
